fix incorrect sunset time
When trying out your script i noticed it was giving a moon(🌙) when it should have given a sun(☀). Turns out when getting the sunrise and sunset times and converting to local timezone it was setting the sunset time in the past due to having a negative offset. This is my simple fix